The 'Sistema_Bancario' code implements a simple banking system using classes in Python. It consists of several classes:
User: Represents a bank user and stores information such as CPF, username, date of birth, address, and password.

Accounts (abstract class): Serves as the base class for different types of bank accounts. It defines abstract methods for deposit, cashout, and statement, which need to be implemented by its subclasses.

Account (subclass of Accounts): Represents a regular bank account. It provides implementations for the abstract methods defined in the Accounts class. It supports depositing and withdrawing money, as well as generating a bank statement.

Storage: Handles the loading and saving of user data to a JSON file. It provides methods to load and save user data from/to the file.

CheckingAccount (subclass of Account): Represents a checking account, which is a specialized type of account. It extends the Account class and adds an overdraft limit, allowing customers to withdraw more money than their account balance (up to a certain limit).

Bank: Acts as the main interface for the banking system. It manages user registration, login, and account creation. It also provides the main menu for banking operations such as deposit, cashout, and statement.

The code utilizes JSON file storage to persist user data, allowing users to register, login, create accounts, and perform banking operations like deposits and withdrawals.

Overall, the code provides a basic framework for a banking system, demonstrating the usage of classes, inheritance, abstract methods, and file storage in Python.